HaiDT commit 17h-10 15/01/2024

+++ medical_supplies/src/main/java/com/example/medical_supplies/service/product/IProductService.java
@@ -9,7 +9,7 @@ public interface IProductService {
 
     Page<Products> findAllProduct(Pageable pageable);
 
-    Products getProductById(Long id);
+    Products getProductById(Integer id);
 
     boolean createProduct(ProductDTO productDTO);
